chore: move tests of radio component from src to tests directory
